backend/requirements.txt
# Async HTTP
aiohttp==3.9.3
asgiref==3.7.2

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0

# LinkedIn Integration
linkedin-api==2.0.3  # Unofficial LinkedIn API wrapper

# Environment & Utils
python-dotenv==1.0.1
urllib3==2.2.0

# FastAPI & Dependencies
fastapi==0.109.2
python-multipart==0.0.9
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25

# Document Processing
pymupdf==1.23.8               # for PDF parsing (fitz)
python-docx==1.1.0            # for DOCX files
pillow==10.2.0               # for image handling
pytesseract==0.3.10          # for OCR

# LangChain & AI
langchain==0.1.0
langchain-google-genai==0.0.7
python-dotenv==1.0.1         # for environment variables

# Database Dependencies
sqlalchemy==2.0.25
aiosqlite==0.19.0      # For async SQLite support

# FastAPI Core
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25

# CORS Support
fastapi-cors==0.0.6

# Database & ORM
sqlalchemy==2.0.25